requirements.txt
# Binance Futures Trading Bot - Dependencies
# Core dependencies
python-binance>=1.0.19
requests>=2.31.0
python-dotenv>=1.0.0

# Data handling (optional - not required for core functionality)
# pandas>=2.0.0
# numpy>=1.24.0

# Async support
aiohttp>=3.9.0

# Logging and monitoring
colorlog>=6.7.0

# Configuration
pyyaml>=6.0

# Testing (optional)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
